Update event loop handling
